Update controllers-hooks-ecosystem.mdx
Line 166:
"This section discusses the communication and inner implementation of Controllers and Hooks. You don't need to read or know it to properly use Cthese blueprints."
Typo corrected to:
"This section discusses the communication and inner implementation of Controllers and Hooks. You don't need to read or know it to properly use these blueprints."