Fix virtual double press event for deconz
This change modifies all helper_last_controller_event updates to force
the trigger_action to be written (and appropriately quoted) as a string.

There are a collection of regexes that test the contents of this json
and expect this value to be quoted.  A better change might be to update
the regexes but this seems safer.