Fix the IKEA E1744
A functional issue with the Home Assistant ZHA blueprint was identified as a result of recent code changes. The root cause of this issue pertains to the modified format of the "args" values received by the blueprint. To address this, the issue has been resolved by retrieving the value of the first element of the "args", which is sufficient to determine the intended action. It should be noted that the "toggle" command does not have any "args". Therefore, an "if" statement has been implemented to handle this case.